move axiosUtil to existing Util folder
